Fix when forcing cuda to be disabled
The GpuOp class must be available even when Cuda is disabled (e.g. by
forcing cuda_available = False). Otherwise some tests fail.

This commit moves the definition of GpuOp directly into cuda/__init__.py
so that it works.
